backend/requirements.txt
Flask==3.0.2
Flask-Cors==5.0.0
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.1.1
psycopg2==2.9.9
python-dotenv==1.0.1
